[20] Move all supervisor start logic to start.go
